requirements.txt
Package                  Version
------------------------ ---------
aiohttp                  3.8.4
aiohttp-retry            2.8.3
aiosignal                1.3.1
alembic                  1.11.1
amqp                     5.1.1
antlr4-python3-runtime   4.9.3
anyio                    3.6.2
app                      0.0.1
appdirs                  1.4.4
argon2-cffi              21.3.0
argon2-cffi-bindings     21.2.0
arrow                    1.2.3
asttokens                2.2.1
async-timeout            4.0.2
asyncssh                 2.13.1
atpublic                 3.1.1
attrs                    23.1.0
backcall                 0.2.0
beautifulsoup4           4.12.2
billiard                 3.6.4.0
bleach                   6.0.0
blinker                  1.6.2
celery                   5.2.7
certifi                  2022.12.7
cffi                     1.15.1
charset-normalizer       3.1.0
click                    8.1.3
click-didyoumean         0.3.0
click-plugins            1.1.1
click-repl               0.2.0
cloudpickle              2.2.1
colorama                 0.4.6
comm                     0.1.3
configobj                5.0.8
contourpy                1.0.7
cryptography             40.0.2
cycler                   0.11.0
databricks-cli           0.17.7
debugpy                  1.6.7
decorator                5.1.1
defusedxml               0.7.1
dictdiffer               0.9.0
diskcache                5.6.1
distro                   1.8.0
docker                   6.1.2
dpath                    2.1.5
dulwich                  0.21.5
dvc                      2.57.2
dvc-data                 0.51.0
dvc-http                 2.30.2
dvc-objects              0.22.0
dvc-render               0.5.3
dvc-studio-client        0.9.2
dvc-task                 0.2.1
entrypoints              0.4
executing                1.2.0
fastapi                  0.95.2
fastjsonschema           2.16.3
filelock                 3.12.0
Flask                    2.3.2
flatten-dict             0.4.2
flufl.lock               7.1.1
fonttools                4.39.3
fqdn                     1.5.1
frozenlist               1.3.3
fsspec                   2023.5.0
funcy                    2.0
gitdb                    4.0.10
GitPython                3.1.31
grandalf                 0.8
greenlet                 2.0.2
h11                      0.14.0
hydra-core               1.3.2
idna                     3.4
imbalanced-learn         0.10.1
imblearn                 0.0
importlib-metadata       6.6.0
ipykernel                6.23.0
ipython                  8.13.2
ipython-genutils         0.2.0
ipython-sql              0.5.0
ipywidgets               8.0.6
isoduration              20.11.0
iterative-telemetry      0.0.8
itsdangerous             2.1.2
jedi                     0.18.2
Jinja2                   3.1.2
joblib                   1.2.0
jsonpointer              2.3
jsonschema               4.17.3
jupyter                  1.0.0
jupyter_client           8.2.0
jupyter-console          6.6.3
jupyter_core             5.3.0
jupyter-events           0.6.3
jupyter_server           2.5.0
jupyter_server_terminals 0.4.4
jupyterlab-pygments      0.2.2
jupyterlab-widgets       3.0.7
kiwisolver               1.4.4
kombu                    5.2.4
Mako                     1.2.4
Markdown                 3.4.3
markdown-it-py           2.2.0
MarkupSafe               2.1.2
matplotlib               3.7.1
matplotlib-inline        0.1.6
mdurl                    0.1.2
mistune                  2.0.5
mlflow                   2.3.2
multidict                6.0.4
nanotime                 0.5.2
nbclassic                1.0.0
nbclient                 0.7.4
nbconvert                7.4.0
nbformat                 5.8.0
nest-asyncio             1.5.6
networkx                 3.1
notebook                 6.5.4
notebook_shim            0.2.3
numpy                    1.24.2
oauthlib                 3.2.2
omegaconf                2.3.0
orjson                   3.8.12
packaging                23.1
pandas                   1.5.3
pandocfilters            1.5.0
parso                    0.8.3
pathspec                 0.11.1
patsy                    0.5.3
pickleshare              0.7.5
Pillow                   9.5.0
pip                      23.0
platformdirs             3.5.0
plotly                   5.14.1
prettytable              3.7.0
prometheus-client        0.16.0
prompt-toolkit           3.0.38
protobuf                 4.23.1
psutil                   5.9.5
psycopg2-binary          2.9.5
pure-eval                0.2.2
py                       1.11.0
pyarrow                  11.0.0
pycparser                2.21
pydantic                 1.10.7
pydot                    1.4.2
pygit2                   1.12.1
Pygments                 2.15.1
pygtrie                  2.5.0
PyJWT                    2.7.0
pyparsing                3.0.9
PyQt5                    5.15.9
PyQt5-Qt5                5.15.2
PyQt5-sip                12.12.1
PyQtWebEngine            5.15.6
PyQtWebEngine-Qt5        5.15.2
pyrsistent               0.19.3
python-dateutil          2.8.2
python-json-logger       2.0.7
pytz                     2023.2
pywin32                  306
pywinpty                 2.0.10
PyYAML                   6.0
pyzmq                    25.0.2
qtconsole                5.4.3
QtPy                     2.3.1
querystring-parser       1.2.4
requests                 2.28.2
retry                    0.9.2
rfc3339-validator        0.1.4
rfc3986-validator        0.1.1
rich                     13.3.5
ruamel.yaml              0.17.26
ruamel.yaml.clib         0.2.7
scikit-learn             1.2.2
scipy                    1.9.3
scmrepo                  1.0.3
seaborn                  0.12.2
Send2Trash               1.8.2
setuptools               67.1.0
shortuuid                1.0.11
shtab                    1.6.1
six                      1.16.0
smmap                    5.0.0
sniffio                  1.3.0
soupsieve                2.4.1
SQLAlchemy               2.0.12
sqlparse                 0.4.4
sqltrie                  0.3.1
stack-data               0.6.2
starlette                0.27.0
statsmodels              0.14.0
tabulate                 0.9.0
tenacity                 8.2.2
terminado                0.17.1
threadpoolctl            3.1.0
tinycss2                 1.2.1
tomlkit                  0.11.8
tornado                  6.3.1
tqdm                     4.65.0
traitlets                5.9.0
typer                    0.9.0
typing_extensions        4.5.0
uri-template             1.2.0
urllib3                  1.26.15
uvicorn                  0.22.0
vine                     5.0.0
voluptuous               0.13.1
waitress                 2.1.2
wcwidth                  0.2.6
webcolors                1.13
webencodings             0.5.1
websocket-client         1.5.1
Werkzeug                 2.3.4
wheel                    0.38.4
widgetsnbextension       4.0.7
xgboost                  1.7.5
yarl                     1.9.2
zc.lockfile              3.0.post1
zipp                     3.15.0
